refactor(structure) prepare structure for merge
